Update 2022-08-02-leaving-facebook-meta-best-for-trino.md
Removed Airbnb entry.

Airbnb is a horrible company that aims to steal money from people via their own "internal" arbitration process.  This allows a host of a place to basically do whatever they want without recourse, include intentional misrepresentation of lodging, amenities, etc., and Airbnb's stance will be to side with the host 100% of the time.  It doesn't seem like a great company, such as Trino.io, would want to associate or indirectly promote this kind of sleazy organization, so removal of all Airbnb-related assets would be in Trino's best interests if they value equitable treatment of people.