Use Python highlighting and change code snippet
The original snippet correctly uses {{ and }} and is missing the Python
highlighting. However that produces a misleading warning when running Jekyll
that could send any contributor down a trap of trying to investigate and fix.
Given that the blog post is old and it still renders fine I added wrong
spaces between the { and } symbols. This makes cut and paste approach invalid
but that is okay now that the post is very old